# api/swagger/v1/todo-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ToDo service",
    "version": "1.0",
    "contact": {
      "name": "go-grpc-http-rest-microservice-tutorial project",
      "url": "https://github.com/amsokol/go-grpc-http-rest-microservice-tutorial",
      "email": "medium@amsokol.com"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/todo/all": {
      "get": {
        "summary": "Read All todo tasks",
        "operationId": "ReadAll",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1ReadAllResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "api",
            "description": "API versioning: it is my best practice to specify version explicitly.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "ToDoService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "v1ReadAllResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "toDos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ToDo"
          },
          "title": "List of all todo tasks"
        }
      },
      "title": "Contains list of all todo tasks"
    },
...
    "v1ToDo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Unique integer identifier of the todo task"
        },
        "title": {
          "type": "string",
          "title": "Title of the task"
        },
        "description": {
          "type": "string",
          "title": "Detail description of the todo task"
        },
        "reminder": {
          "type": "string",
          "format": "date-time",
          "title": "Date and time to remind the todo task"
        }
      },
      "title": "Task we have to do"
    },
...
  }
}